fclShop FUN_003EE920 MATCH (plain sin/cos + reordered alias), FUN_003EF2B0 nd1003->995, FUN_003E8860 nd953->952; bp_tex FUN_00221B60 nd315->167 (switch-goto dispatch); k_fldFrame func_001ac950 nd69->59

+++ src/Yajima/y_runtime.c
@@ -8533,7 +8533,7 @@ short FUN_0043c180(char param_1)
   int iVar10;
   short sVar9;
   u16 uVar8;
-  int player;
+  u8 player;
   u8 *row;
   sVar9 = 0;
   uVar8 = 0;
